+++ mmu2-diy/axis.cpp
@@ -1,0 +1,7 @@
+#include "axis.h"
+
+Axis::Axis(PinNr enable, PinNr direcetion, PinNr step, uint16_t fullStepsPerRev, uint16_t resolution, uint16_t maxFullSteps)
+{
+
+}
+