+++ mmu2-diy/config.h
@@ -1,7 +1,11 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
+#include <stdint.h>
 
+#define MMU2_VERSION "0.5.0  2018-10-18"
+#define FW_VERSION 50             // config.h  (MM-control-01 firmware)
+#define FW_BUILDNR 85             // config.h  (MM-control-01 firmware)
 
 //#define DEBUG
 //#define DEBUGIDLER
@@ -9,8 +13,12 @@
 //#define CRAZYIVAN
 //#define TURNOFFSELECTORMOTOR
 
-#define DIY_BOARD
+//#define DIY_BOARD
 #define PRUSA_BOARD
 
 #endif // CONFIG_H
+
+
+
+