+++ mmu2-diy/shiftregister.cpp
@@ -1,6 +1,6 @@
 #include "shiftregister.h"
 
-ShiftRegister::ShiftRegister(uint8_t nrOfBits)
+ShiftRegister::ShiftRegister(uint8_t nrOfBits, PinNr data, PinNr clock, PinNr latch)
 	:N(nrOfBits)
 {
 	if(N > 32){
@@ -11,5 +11,14 @@ ShiftRegister::ShiftRegister(uint8_t nrOfBits)
 	}
 
 	data = 0; // clear all bits by default
+}
+
+void ShiftRegister::writeBit(uint8_t bitNr, bool value)
+{
+	if(value){
+		data |= (1<<bitNr);
+	}else{
+		data &= ~(1<<bitNr);
+	}
 }