+++ mmu2-diy/shiftregister.h
@@ -2,11 +2,12 @@
 #define SHIFTREGISTER_H
 
 #include <stdint.h>
+#include "defs.h"
 
 class ShiftRegister
 {
 public:
-	ShiftRegister(uint8_t nrOfBits);
+	ShiftRegister(uint8_t nrOfBits, PinNr data, PinNr clock, PinNr latch);
 
 	/**
 	 * @brief setBits
@@ -23,22 +24,33 @@ public:
 	void clearBits(uint32_t mask);
 
 	/**
-	 * @brief setData
+	 * @brief writeData
 	 * overwrite current data
 	 * @param data
 	 */
-	void setData(uint32_t data);
+	void writeData(uint32_t data);
 
 	/**
-	 * @brief writeData
+	 * @brief transferData
 	 * clocks current data out to the shift registers.
 	 * MSB first.
 	 */
-	void writeData();
+	void transferData();
+
+	/**
+	 * @brief writeBit
+	 * Writes a single bit to the data in RAM
+	 * @param bitNr starting from 0 until N-1
+	 * @param value: 0 or 1
+	 */
+	void writeBit(uint8_t bitNr, bool value);
 
 	/// Number of bits, usually 8, 16, 24 or 32 (max.)
 	uint8_t N;
 	uint32_t data;
+	PinNr dataPin;
+	PinNr clockPin;
+	PinNr latchPin;
 };
 
 #endif // SHIFTREGISTER_H